arch/x86_64-all: Make LIB_VECTSIZE an (int) instead of (size_t)
size_t is a long long on x86_64, and when multiplied by negative
numbers, results in a value that causes compiler complaints for
Exec/SetFunction().

Since LIB_VECTSIZE is small, we'll just cast it to an (int) here, and
everything is now happy.


git-svn-id: https://svn.aros.org/svn/aros/trunk@39362 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/x86_64-all/include/aros/cpu.h
@@ -101,7 +101,7 @@ struct JumpVec
 };
 
 /* Use these to access a vector table */
-#define LIB_VECTSIZE			(sizeof (struct JumpVec))
+#define LIB_VECTSIZE			((int)sizeof (struct JumpVec))
 #define __AROS_GETJUMPVEC(lib,n)        (&((struct JumpVec *)lib)[-(long)(n)])
 #define __AROS_GETVECADDR(lib,n)        (__AROS_GETJUMPVEC(lib,n)->vec)
 #define __AROS_SETVECADDR(lib,n,addr)   (__AROS_GETJUMPVEC(lib,n)->vec = (addr))
